Incorporate canvas copying so diff sizes work.
Now the sprite's image is copied to a canvas as big as the largest width and height. The new image is copy to its own large canvas. Then the canvases are morphed. Finally, the sprite's image is replaced with the new image, to get rid of extra transparent pixels which could affect wall collisions.